+++ src/SharpCode/Code.cs
@@ -1,4 +1,5 @@
 using System;
+using Optional;
 
 namespace SharpCode
 {
@@ -22,6 +23,49 @@ namespace SharpCode
         /// </param>
         public static NamespaceBuilder CreateNamespace(string name) =>
             new NamespaceBuilder(name);
+
+        /// <summary>
+        /// Creates a new <see cref="EnumBuilder"/> instance for building enums.
+        /// </summary>
+        public static EnumBuilder CreateEnum() =>
+            new EnumBuilder();
+
+        /// <summary>
+        /// Creates a new pre-configured <see cref="EnumBuilder"/> instance for building enums. Configures the
+        /// <see cref="EnumBuilder"/> instance with the specified <paramref name="name"/> and
+        /// <paramref name="accessModifier"/>.
+        /// </summary>
+        /// <param name="name">
+        /// The name of the enum.
+        /// </param>
+        /// <param name="accessModifier">
+        /// The access modifier of the enum.
+        /// </param>
+        public static EnumBuilder CreateEnum(string name, AccessModifier accessModifier = AccessModifier.Public) =>
+            new EnumBuilder(name, accessModifier);
+
+        /// <summary>
+        /// Creates a new <see cref="EnumMemberBuilder"/> instance for building enum members.
+        /// </summary>
+        public static EnumMemberBuilder CreateEnumMember() =>
+            new EnumMemberBuilder();
+
+        /// <summary>
+        /// Creates a new pre-configured <see cref="EnumMemberBuilder"/> instance for building enum members. Configures
+        /// the <see cref="EnumMemberBuilder"/> instance with the specified <paramref name="name"/> and
+        /// <paramref name="value"/>.
+        /// </summary>
+        /// <param name="name">
+        /// The name of the enum member.
+        /// </param>
+        /// <param name="value">
+        /// Optional. Explicitly specifies the value of the enum member. <b>Note</b> that if the enum is marked as
+        /// flags, via <see cref="EnumBuilder.MakeFlagsEnum(bool)"/>, you need to explicitly provide the values of all
+        /// members to ensure correct functionality. Flags enums for which no member has an explicit value will
+        /// auto-generate appropriate values for each member.
+        /// </param>
+        public static EnumMemberBuilder CreateEnumMember(string name, int? value = null) =>
+            new EnumMemberBuilder(name, value.ToOption());
 
         /// <summary>
         /// Creates a new <see cref="InterfaceBuilder"/> instance for building interface structures.