+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Automated GitHub releases from the release action
+- Support for enums generation
+- Support for various enum members - with implicit and/or explicit values, flags
 
 ## [0.1.0] - 2020-09-26